Sophon improvements
+ Separate Sophon methods to its own files
+ Add mechanism to check remained disk space on Preload
