+++ CollapseLauncher/Classes/InstallManagement/BaseClass/InstallManagerBase.cs
@@ -30,9 +30,6 @@ using Hi3Helper.Http.Legacy;
 using Hi3Helper.SentryHelper;
 using Hi3Helper.Shared.ClassStruct;
 using Hi3Helper.Shared.Region;
-using Hi3Helper.Sophon;
-using Hi3Helper.Sophon.Infos;
-using Hi3Helper.Sophon.Structs;
 using Microsoft.UI.Text;
 using Microsoft.UI.Xaml;
 using Microsoft.UI.Xaml.Controls;
@@ -41,7 +38,6 @@ using SevenZipExtractor;
 using SharpHDiffPatch.Core;
 using SharpHDiffPatch.Core.Event;
 using System;
-using System.Collections.Concurrent;
 using System.Collections.Generic;
 using System.Diagnostics;
 using System.Diagnostics.CodeAnalysis;
@@ -64,9 +60,6 @@ using CoreCombinedStream = Hi3Helper.EncTool.CombinedStream;
 using ZipArchive = SharpCompress.Archives.Zip.ZipArchive;
 using ZipArchiveEntry = SharpCompress.Archives.Zip.ZipArchiveEntry;
 #endif
-
-using SophonLogger = Hi3Helper.Sophon.Helper.Logger;
-using SophonManifest = Hi3Helper.Sophon.SophonManifest;
 
 // ReSharper disable ForCanBeConvertedToForeach
 // ReSharper disable SwitchStatementHandlesSomeKnownEnumValuesWithDefault
@@ -136,7 +129,6 @@ namespace CollapseLauncher.InstallManager.Base
                          $"{Path.GetFileNameWithoutExtension(_gameVersionManager.GamePreset.GameExecutableName)}_Data");
 
         protected virtual string _gameDataPersistentPath => Path.Combine(_gameDataPath, "Persistent");
-        protected virtual string _gameSophonChunkDir => Path.Combine(_gamePath, "chunk_collapse");
         protected virtual string _gameAudioLangListPath => null;
         protected virtual string _gameAudioLangListPathStatic => null;
         protected IRepair _gameRepairTool { get; set; }
@@ -154,40 +146,6 @@ namespace CollapseLauncher.InstallManager.Base
         private   long                     _totalLastSizeCurrent;
 
         protected bool _isAllowExtractCorruptZip  { get; set; }
-        protected bool _isSophonDownloadCompleted { get; set; }
-
-        protected bool _isSophonPreloadCompleted
-        {
-            get => File.Exists(Path.Combine(_gameSophonChunkDir, PreloadVerifiedFileName));
-            set
-            {
-                string verifiedFile =
-                    EnsureCreationOfDirectory(Path.Combine(_gameSophonChunkDir, PreloadVerifiedFileName));
-                try
-                {
-                    FileInfo fileInfo = new FileInfo(verifiedFile);
-                    if (value)
-                    {
-                        fileInfo.Create().Dispose();
-                        return;
-                    }
-
-                    if (fileInfo.Exists)
-                    {
-                        fileInfo.IsReadOnly = false;
-                        fileInfo.Delete();
-                    }
-                }
-                catch (Exception ex)
-                {
-                    SentryHelper.ExceptionHandler(ex, SentryHelper.ExceptionType.UnhandledOther);
-                    LogWriteLine($"Error while deleting/creating sophon preload completion file! {ex}", LogType.Warning,
-                                 true);
-                }
-            }
-        }
-
-        protected List<string>           _sophonVOLanguageList  { get; set; } = [];
         protected UninstallGameProperty? _uninstallGameProperty { get; set; }
 
         #endregion
@@ -197,38 +155,6 @@ namespace CollapseLauncher.InstallManager.Base
         public event EventHandler FlushingTrigger;
         public virtual bool       StartAfterInstall { get; set; }
         public virtual bool       IsRunning         { get; protected set; }
-
-        public virtual bool IsUseSophon =>
-            _gameVersionManager.GamePreset.LauncherResourceChunksURL != null
-            && !File.Exists(Path.Combine(_gamePath, "@DisableSophon"))
-            && !_canDeltaPatch && !_forceIgnoreDeltaPatch
-            && LauncherConfig.GetAppConfigValue("IsEnableSophon").ToBool();
-
-        protected virtual int SophonGetThreadNum()
-        {
-            // Get from config
-            var n = LauncherConfig.GetAppConfigValue("SophonCpuThread").ToInt();
-            if (n == 0) // If config is default "0", then use sqrt of thread number as safe number
-            {
-                n = (int)Math.Sqrt(Environment.ProcessorCount);
-            }
-
-            return Math.Clamp(n, 2, 64); // Clamp value to prevent errors
-        }
-
-        protected virtual int SophonGetHttpHandler()
-        {
-            var n = LauncherConfig.GetAppConfigValue("SophonHttpConnInt").ToInt();
-            if (n == 0)
-            {
-                n = (int)Math.Sqrt(Environment.ProcessorCount) * 2;
-            }
-
-            return Math.Clamp(n, 4, 128);
-        }
-
-        public virtual bool IsSophonInUpdateMode => _isSophonInUpdateMode;
-
         #endregion
 
         public InstallManagerBase(UIElement parentUI, IGameVersionCheck GameVersionManager)
@@ -785,620 +711,6 @@ namespace CollapseLauncher.InstallManager.Base
             return 1;
         }
 
-        public virtual async Task StartPackageInstallSophon(GameInstallStateEnum gameState)
-        {
-            // Set the flag to false
-            _isSophonDownloadCompleted = false;
-
-            // Set the max thread and httpHandler based on settings
-            int maxThread       = SophonGetThreadNum();
-            int maxChunksThread = Math.Clamp(maxThread / 2, 2, 32);
-            int maxHttpHandler  = Math.Max(maxThread, SophonGetHttpHandler());
-
-            LogWriteLine($"Initializing Sophon Chunk download method with Main Thread: {maxThread}, Chunks Thread: {maxChunksThread} and Max HTTP handle: {maxHttpHandler}",
-                         LogType.Default, true);
-
-            // Initialize the HTTP client
-            HttpClient httpClient = new HttpClientBuilder()
-                .UseLauncherConfig(maxHttpHandler)
-                .Create();
-
-            using (ThreadPoolThrottle.Start())
-            {
-                try
-                {
-                    // Reset status and progress properties
-                    ResetStatusAndProgress();
-
-                    // Clear the VO language list
-                    _sophonVOLanguageList?.Clear();
-
-                    // Subscribe the logger event
-                    SophonLogger.LogHandler += UpdateSophonLogHandler;
-
-                    // Get the requested URL and version based on current state.
-                    if (_gameVersionManager.GamePreset
-                                           .LauncherResourceChunksURL != null)
-                    {
-#nullable enable
-                        // Reassociate the URL if branch url exist
-                        string? branchUrl = _gameVersionManager.GamePreset
-                                                               .LauncherResourceChunksURL
-                                                               .BranchUrl;
-                        if (!string.IsNullOrEmpty(branchUrl)
-                            && !string.IsNullOrEmpty(_gameVersionManager.GamePreset.LauncherBizName))
-                        {
-                            await _gameVersionManager.GamePreset
-                                                     .LauncherResourceChunksURL
-                                                     .EnsureReassociated(
-                                                                         httpClient,
-                                                                         branchUrl,
-                                                                         _gameVersionManager.GamePreset.LauncherBizName,
-                                                                         _token.Token);
-                        }
-#nullable restore
-
-#if SIMULATEAPPLYPRELOAD
-                         string requestedUrl = gameState switch
-                         {
-                             GameInstallStateEnum.InstalledHavePreload => _gameVersionManager.GamePreset
-                                .LauncherResourceChunksURL.PreloadUrl,
-                             _ => _gameVersionManager.GamePreset.LauncherResourceChunksURL.MainUrl
-                         };
-                         GameVersion? requestedVersion = gameState switch
-                         {
-                             GameInstallStateEnum.InstalledHavePreload => _gameVersionManager!
-                                .GetGameVersionAPIPreload(),
-                             _ => _gameVersionManager!.GetGameVersionAPIPreload()
-                         } ?? _gameVersionManager!.GetGameVersionAPI();
-#else
-                        string requestedUrl = gameState switch
-                        {
-                            GameInstallStateEnum.InstalledHavePreload => _gameVersionManager
-                               .GamePreset
-                               .LauncherResourceChunksURL.PreloadUrl,
-                            _ => _gameVersionManager.GamePreset.LauncherResourceChunksURL.MainUrl
-                        };
-                        GameVersion? requestedVersion = gameState switch
-                        {
-                            GameInstallStateEnum.InstalledHavePreload =>
-                                _gameVersionManager!
-                                   .GetGameVersionAPIPreload(),
-                            _ => _gameVersionManager!.GetGameVersionAPI()
-                        } ?? _gameVersionManager!.GetGameVersionAPI();
-
-                        // Add the tag query to the Url
-                        requestedUrl += $"&tag={requestedVersion.ToString()}";
-#endif
-
-                        // Set the progress bar to indetermined
-                        _status.IsIncludePerFileIndicator = false;
-                        _status.IsProgressPerFileIndetermined = false;
-                        _status.IsProgressAllIndetermined = true;
-                        UpdateStatus();
-
-                        // Initialize the info pair list
-                        var sophonInfoPairList = new List<SophonChunkManifestInfoPair>();
-
-                        // Get the info pair based on info provided above (for main game file)
-                        var sophonMainInfoPair = await
-                            SophonManifest.CreateSophonChunkManifestInfoPair(httpClient, requestedUrl, "game", _token.Token);
-
-                        // Ensure that the manifest is ordered based on _gameVoiceLanguageLocaleIdOrdered
-                        RearrangeSophonDataLocaleOrder(sophonMainInfoPair.OtherSophonData);
-
-                        // Add the manifest to the pair list
-                        sophonInfoPairList.Add(sophonMainInfoPair);
-
-                        List<string> voLanguageList =
-                            GetSophonLanguageDisplayDictFromVoicePackList(sophonMainInfoPair.OtherSophonData);
-
-                        // Get Audio Choices first
-                        (List<int> addedVO, int setAsDefaultVO) =
-                            await Dialog_ChooseAudioLanguageChoice(voLanguageList, GetSophonLocaleCodeIndex(sophonMainInfoPair.OtherSophonData, "ja-jp"));
-
-                        try
-                        {
-                            if (addedVO == null || setAsDefaultVO < 0)
-                            {
-                                throw new TaskCanceledException();
-                            }
-
-                            for (int i = 0; i < addedVO.Count; i++)
-                            {
-                                int voLangIndex = addedVO[i];
-                                string voLangLocaleCode = GetLanguageLocaleCodeByID(voLangIndex);
-                                _sophonVOLanguageList?.Add(voLangLocaleCode);
-
-                                // Get the info pair based on info provided above (for the selected VO audio file)
-                                SophonChunkManifestInfoPair sophonSelectedVoLang =
-                                    sophonMainInfoPair.GetOtherManifestInfoPair(voLangLocaleCode);
-                                sophonInfoPairList.Add(sophonSelectedVoLang);
-                            }
-
-                            // Set the voice language ID to value given
-                            _gameVersionManager.GamePreset.SetVoiceLanguageID(setAsDefaultVO);
-
-                            // Get the remote total size and current total size
-                            _progressAllCountTotal = sophonInfoPairList.Sum(x => x.ChunksInfo.FilesCount);
-                            _progressAllSizeTotal = sophonInfoPairList.Sum(x => x.ChunksInfo.TotalSize);
-                            _progressAllSizeCurrent = 0;
-
-                            // Set the display to Install Mode
-                            _isSophonInUpdateMode = false;
-
-                            // Set the progress bar to indetermined
-                            _status.IsIncludePerFileIndicator = false;
-                            _status.IsProgressPerFileIndetermined = false;
-                            _status.IsProgressAllIndetermined = false;
-                            UpdateStatus();
-                        }
-                        catch (TaskCanceledException)
-                        {
-                            throw;
-                        }
-                        catch (OperationCanceledException)
-                        {
-                            throw;
-                        }
-                        catch (Exception e)
-                        {
-                            ErrorSender.SendException(e);
-                        }
-
-                        // Get the parallel options
-                        var parallelOptions = new ParallelOptions
-                        {
-                            MaxDegreeOfParallelism = maxThread,
-                            CancellationToken = _token.Token
-                        };
-                        var parallelChunksOptions = new ParallelOptions
-                        {
-                            MaxDegreeOfParallelism = maxChunksThread,
-                            CancellationToken = _token.Token
-                        };
-
-                        // Declare the download delegate
-                        async ValueTask DelegateAssetDownload(SophonAsset asset, CancellationToken _)
-                        {
-                            // ReSharper disable once AccessToDisposedClosure
-                            await RunSophonAssetDownloadThread(httpClient, asset, parallelChunksOptions);
-                        }
-
-                        // Declare the rename temp file delegate
-                        async ValueTask DelegateAssetRenameTempFile(SophonAsset asset, CancellationToken token)
-                        {
-                            await Task.Run(() =>
-                            {
-                                // If the asset is a dictionary, then return
-                                if (asset.IsDirectory)
-                                {
-                                    return;
-                                }
-
-                                // Get the file path and start the write process
-                                var assetName = asset.AssetName;
-                                var filePath = new FileInfo(
-                                    EnsureCreationOfDirectory(Path.Combine(_gamePath, assetName)) +
-                                    "_tempSophon").EnsureNoReadOnly();
-                                var origFilePath = new FileInfo(Path.Combine(_gamePath, assetName)).EnsureNoReadOnly(out bool isExist);
-
-                                if (isExist)
-                                {
-                                    filePath.MoveTo(origFilePath.FullName, true);
-                                    filePath.Refresh();
-                                    origFilePath.Refresh();
-                                }
-                            }, token);
-                        }
-
-                        // Enumerate the asset in parallel and start the download process
-                        await RunTaskAction(httpClient, sophonInfoPairList, parallelOptions, DelegateAssetDownload);
-
-                        // Rename temporary files
-                        await RunTaskAction(httpClient, sophonInfoPairList, parallelOptions, DelegateAssetRenameTempFile);
-
-                        // Remove sophon verified files
-                        CleanupTempSophonVerifiedFiles();
-                    }
-
-                    _isSophonDownloadCompleted = true;
-                }
-                finally
-                {
-                    // Unsubscribe the logger event
-                    SophonLogger.LogHandler -= UpdateSophonLogHandler;
-                    httpClient.Dispose();
-                }
-            }
-
-            return;
-
-            async Task RunTaskAction(HttpClient client, List<SophonChunkManifestInfoPair> sophonInfoPairListLocal,
-                                     ParallelOptions parallelOptions,
-                                     Func<SophonAsset, CancellationToken, ValueTask> actionDelegate)
-            {
-                // Create a sophon download speed limiter instance
-                SophonDownloadSpeedLimiter downloadSpeedLimiter = SophonDownloadSpeedLimiter.CreateInstance(LauncherConfig.DownloadSpeedLimitCached);
-
-                try
-                {
-                    LauncherConfig.DownloadSpeedLimitChanged += downloadSpeedLimiter.GetListener();
-                    var processingInfoPair = new ConcurrentDictionary<SophonChunksInfo, byte>();
-                    var infoPairListCopy   = sophonInfoPairListLocal.ToList();
-                    foreach (SophonChunkManifestInfoPair sophonDownloadInfoPair in infoPairListCopy)
-                    {
-                        if (!processingInfoPair.TryAdd(sophonDownloadInfoPair.ChunksInfo, 0))
-                        {
-                            LogWriteLine($"Found duplicate operation for {sophonDownloadInfoPair.ChunksInfo.ChunksBaseUrl}! Skipping...",
-                                         LogType.Warning, true);
-                            continue;
-                        }
-                        // Enumerate in parallel and process the assets
-                        await Parallel.ForEachAsync(SophonManifest.EnumerateAsync(client, sophonDownloadInfoPair,
-                                                                                  downloadSpeedLimiter),
-                                                    parallelOptions,
-                                                    actionDelegate).ConfigureAwait(false);
-                        processingInfoPair.Remove(sophonDownloadInfoPair.ChunksInfo, out _);
-                    }
-                }
-                finally
-                {
-                    LauncherConfig.DownloadSpeedLimitChanged -= downloadSpeedLimiter.GetListener();
-                }
-            }
-        }
-
-        public virtual async Task StartPackageUpdateSophon(GameInstallStateEnum gameState, bool isPreloadMode)
-        {
-            // Set the flag to false
-            _isSophonDownloadCompleted = false;
-
-            // Set the max thread and httpHandler based on settings
-            int maxThread       = SophonGetThreadNum();
-            int maxChunksThread = Math.Clamp(maxThread / 2, 2, 32);
-            int maxHttpHandler  = Math.Max(maxThread, SophonGetHttpHandler());
-
-            LogWriteLine($"Initializing Sophon Chunk update method with Main Thread: {maxThread}, Chunks Thread: {maxChunksThread} and Max HTTP handle: {maxHttpHandler}",
-                         LogType.Default, true);
-
-            // Initialize the HTTP client
-            HttpClient httpClient = new HttpClientBuilder()
-                .UseLauncherConfig(maxHttpHandler)
-                .Create();
-
-            try
-            {
-                // Reset status and progress properties
-                ResetStatusAndProgress();
-
-                // Clear the VO language list
-                _sophonVOLanguageList?.Clear();
-
-                // Subscribe the logger event
-                SophonLogger.LogHandler += UpdateSophonLogHandler;
-
-                // Init asset list
-                List<SophonAsset> sophonUpdateAssetList = [];
-
-                // Get the previous version details of the preload or the recent update.
-                GameVersion? requestedVersionFrom = _gameVersionManager!.GetGameExistingVersion();
-                if (_gameVersionManager.GamePreset.LauncherResourceChunksURL != null)
-                {
-                #nullable enable
-                    // Reassociate the URL if branch url exist
-                    string? branchUrl = _gameVersionManager.GamePreset
-                                                           .LauncherResourceChunksURL
-                                                           .BranchUrl;
-                    if (!string.IsNullOrEmpty(branchUrl)
-                        && !string.IsNullOrEmpty(_gameVersionManager.GamePreset.LauncherBizName))
-                    {
-                        await _gameVersionManager.GamePreset
-                                                 .LauncherResourceChunksURL
-                                                 .EnsureReassociated(
-                                                                     httpClient,
-                                                                     branchUrl,
-                                                                     _gameVersionManager.GamePreset.LauncherBizName,
-                                                                     _token.Token);
-                    }
-                #nullable restore
-
-                    string requestedBaseUrlFrom = isPreloadMode
-                        ? _gameVersionManager.GamePreset.LauncherResourceChunksURL.PreloadUrl
-                        : _gameVersionManager.GamePreset.LauncherResourceChunksURL.MainUrl;
-                #if SIMULATEAPPLYPRELOAD
-                    string requestedBaseUrlTo = _gameVersionManager.GamePreset.LauncherResourceChunksURL.PreloadUrl;
-                #else
-                    string requestedBaseUrlTo = requestedBaseUrlFrom;
-                #endif
-                    // Add the tag query to the previous version's Url
-                    requestedBaseUrlFrom += $"&tag={requestedVersionFrom.ToString()}";
-
-                    // Create a sophon download speed limiter instance
-                    SophonDownloadSpeedLimiter downloadSpeedLimiter = SophonDownloadSpeedLimiter.CreateInstance(LauncherConfig.DownloadSpeedLimitCached);
-
-                    // Add base game diff data
-                    await AddSophonDiffAssetsToList(httpClient,            requestedBaseUrlFrom, requestedBaseUrlTo,
-                                                    sophonUpdateAssetList, "game",               downloadSpeedLimiter);
-
-                    // If the game has lang list path, then add it
-                    if (_gameAudioLangListPath != null)
-                    {
-                        // Add existing voice-over diff data
-                        await AddSophonAdditionalVODiffAssetsToList(httpClient,         requestedBaseUrlFrom,
-                                                                    requestedBaseUrlTo!, sophonUpdateAssetList,
-                                                                    downloadSpeedLimiter);
-                    }
-                }
-
-                // Get the remote chunk size
-                _progressPerFileSizeTotal   = sophonUpdateAssetList.GetCalculatedDiffSize(!isPreloadMode);
-                _progressPerFileSizeCurrent = 0;
-
-                // Get the remote total size and current total size
-                _progressAllCountTotal = sophonUpdateAssetList.Count(x => !x.IsDirectory);
-                _progressAllSizeTotal = !isPreloadMode
-                    ? sophonUpdateAssetList.Sum(x => x.AssetSize)
-                    : _progressPerFileSizeTotal;
-                _progressAllSizeCurrent = 0;
-
-                // Get the parallel options
-                var parallelOptions = new ParallelOptions
-                {
-                    MaxDegreeOfParallelism = maxThread,
-                    CancellationToken      = _token.Token
-                };
-                var parallelChunksOptions = new ParallelOptions
-                {
-                    MaxDegreeOfParallelism = maxChunksThread,
-                    CancellationToken      = _token.Token
-                };
-
-                // Set the progress bar to indetermined
-                _isSophonInUpdateMode                  = !isPreloadMode;
-                _status.IsIncludePerFileIndicator     = !isPreloadMode;
-                _status.IsProgressPerFileIndetermined = false;
-                _status.IsProgressAllIndetermined     = false;
-                _status.ActivityStatus = $"{(_isSophonInUpdateMode && !isPreloadMode
-                    ? Lang._Misc.UpdatingAndApplying
-                    : Lang._Misc.Downloading)}: {string.Format(Lang._Misc.PerFromTo, _progressAllCountCurrent,
-                                                               _progressAllCountTotal)}";
-                UpdateStatus();
-
-                // Get the update source and destination, also where the staging chunk files will be stored
-                string chunkPath = _gameSophonChunkDir;
-                string gamePath  = _gamePath;
-
-                // If the chunk directory is not exist, then create one.
-                if (!Directory.Exists(chunkPath) && chunkPath != null)
-                {
-                    Directory.CreateDirectory(chunkPath);
-                }
-
-                bool canDeleteChunks          = _canDeleteZip;
-                bool isSophonPreloadCompleted = _isSophonPreloadCompleted;
-
-                // If preload completed and perf mode is on, use all CPU cores 
-                if (isSophonPreloadCompleted && LauncherConfig.GetAppConfigValue("SophonPreloadApplyPerfMode").ToBool())
-                {
-                    maxThread       = Environment.ProcessorCount;
-                    maxChunksThread = Math.Clamp(maxThread / 2, 2, 32);
-                    
-                    parallelOptions = new ParallelOptions
-                    {
-                        MaxDegreeOfParallelism = maxThread,
-                        CancellationToken      = _token.Token
-                    };
-                    parallelChunksOptions = new ParallelOptions
-                    {
-                        MaxDegreeOfParallelism = maxChunksThread,
-                        CancellationToken      = _token.Token
-                    };
-                }
-
-                var processingAsset = new ConcurrentDictionary<SophonAsset, byte>();
-                // Set the delegate function for the download action
-                async ValueTask Action(SophonAsset asset, CancellationToken ctx)
-                {
-                    if (!processingAsset.TryAdd(asset, 0))
-                    {
-                        LogWriteLine($"Found duplicate operation for {asset.AssetName}! Skipping...",
-                                     LogType.Warning, true);
-                        return;
-                    }
-                    if (isPreloadMode)
-                    {
-                        // If preload mode, then only download the chunks
-                        await asset.DownloadDiffChunksAsync(httpClient, chunkPath, parallelChunksOptions,
-                                                            UpdateSophonFileTotalProgress, null,
-                                                            UpdateSophonDownloadStatus, isSophonPreloadCompleted);
-                        return;
-                    }
-
-                    // Ensure to remove the read-only attribute
-                    string currentAssetPath = Path.Combine(gamePath, asset.AssetName);
-                    TryUnassignReadOnlyFileSingle(currentAssetPath);
-
-                    // Otherwise, start the patching process
-                    await asset.WriteUpdateAsync(httpClient, gamePath, gamePath, chunkPath, canDeleteChunks,
-                                                 parallelChunksOptions, UpdateSophonFileTotalProgress,
-                                                 UpdateSophonFileDownloadProgress, UpdateSophonDownloadStatus);
-                    processingAsset.Remove(asset, out _);
-                }
-
-                // Enumerate in parallel and process the assets
-                await Parallel.ForEachAsync(sophonUpdateAssetList.Where(x => !x.IsDirectory),
-                                            parallelOptions,
-                                            Action);
-
-                _isSophonPreloadCompleted = isPreloadMode;
-
-                // If it's in update mode, then clean up the temp sophon verified files
-                if (!isPreloadMode)
-                {
-                    CleanupTempSophonVerifiedFiles();
-                }
-
-                _isSophonDownloadCompleted = true;
-            }
-            finally
-            {
-                // Unsubscribe the logger event
-                SophonLogger.LogHandler -= UpdateSophonLogHandler;
-                httpClient.Dispose();
-
-                // Check if the DXSETUP file is exist, then delete it.
-                // The DXSETUP files causes some false positive detection of data modification
-                // for some games (like Genshin, which causes 4302-x errors for some reason)
-                string dxSetupDir = Path.Combine(_gamePath, "DXSETUP");
-                TryDeleteReadOnlyDir(dxSetupDir);
-            }
-        }
-
-        protected virtual void CleanupTempSophonVerifiedFiles()
-        {
-            DirectoryInfo dirPath = new (_gameSophonChunkDir);
-            try
-            {
-                if (!dirPath.Exists)
-                    return;
-
-                foreach (FileInfo file in dirPath.EnumerateFiles("*.verified", SearchOption.TopDirectoryOnly)
-                    .EnumerateNoReadOnly())
-                {
-                    file.Delete();
-                }
-            }
-            catch
-            {
-                // ignored
-            }
-        }
-
-        private async Task AddSophonDiffAssetsToList(HttpClient                 httpClient,
-                                                     string                     requestedUrlFrom,
-                                                     string                     requestedUrlTo,
-                                                     List<SophonAsset>          sophonPreloadAssetList,
-                                                     string                     matchingField,
-                                                     SophonDownloadSpeedLimiter downloadSpeedLimiter)
-        {
-            // Get the manifest pair for both previous (from) and next (to) version
-            SophonChunkManifestInfoPair requestPairFrom = await SophonManifest
-               .CreateSophonChunkManifestInfoPair(httpClient, requestedUrlFrom, matchingField, _token.Token);
-            SophonChunkManifestInfoPair requestPairTo = await SophonManifest
-               .CreateSophonChunkManifestInfoPair(httpClient, requestedUrlTo, matchingField, _token.Token);
-
-            // Ensure that the manifest is ordered based on _gameVoiceLanguageLocaleIdOrdered
-            RearrangeSophonDataLocaleOrder(requestPairFrom.OtherSophonData);
-            RearrangeSophonDataLocaleOrder(requestPairTo.OtherSophonData);
-
-            // Add asset to the list
-            await foreach (SophonAsset sophonAsset in SophonUpdate
-                                                     .EnumerateUpdateAsync(httpClient, requestPairFrom, requestPairTo,
-                                                                           false, downloadSpeedLimiter)
-                                                     .WithCancellation(_token.Token))
-            {
-                sophonPreloadAssetList.Add(sophonAsset);
-            }
-        }
-
-#nullable enable
-        private async Task AddSophonAdditionalVODiffAssetsToList(HttpClient httpClient,
-                                                         string requestedUrlFrom,
-                                                         string requestedUrlTo,
-                                                         List<SophonAsset> sophonPreloadAssetList,
-                                                         SophonDownloadSpeedLimiter downloadSpeedLimiter)
-        {
-            // Get the main VO language name from Id
-            string mainLangId = GetLanguageLocaleCodeByID(_gameVoiceLanguageID);
-            // Get the manifest pair for both previous (from) and next (to) version for the main VO
-            await AddSophonDiffAssetsToList(httpClient, requestedUrlFrom, requestedUrlTo,
-                                            sophonPreloadAssetList, mainLangId, downloadSpeedLimiter);
-
-            // Check if the audio lang list file is exist, then try add others
-            FileInfo fileInfo = new FileInfo(_gameAudioLangListPath).EnsureNoReadOnly();
-            if (fileInfo.Exists)
-            {
-                // Use stream reader to read the list one-by-one
-                using StreamReader reader = new StreamReader(_gameAudioLangListPath);
-                // Read until EOF
-                string? line;
-                while ((line = await reader.ReadLineAsync()) != null)
-                {
-                    // Get other lang Id, pass it and try add to the list
-                    string? otherLangId = GetLanguageLocaleCodeByLanguageString(line
-#if !DEBUG
-                        , false
-#endif
-                        );
-
-                    // Check if the voice pack is actually the same as default.
-                    if (!string.IsNullOrEmpty(otherLangId) && otherLangId.Equals(mainLangId, StringComparison.OrdinalIgnoreCase))
-                    {
-                        continue;
-                    }
-
-                    // Get the manifest pair for both previous (from) and next (to) version for other VOs
-                    await AddSophonDiffAssetsToList(httpClient, requestedUrlFrom, requestedUrlTo,
-                                                    sophonPreloadAssetList, otherLangId, downloadSpeedLimiter);
-                }
-            }
-        }
-#nullable restore
-
-        private async ValueTask RunSophonAssetDownloadThread(HttpClient      client, SophonAsset asset,
-                                                             ParallelOptions parallelOptions)
-        {
-            // If the asset is a dictionary, then return
-            if (asset.IsDirectory)
-            {
-                return;
-            }
-
-            // Get the file path and start the write process
-            var assetName = asset.AssetName;
-            var filePath  = EnsureCreationOfDirectory(Path.Combine(_gamePath, assetName));
-
-            // Get the target and temp file info
-            FileInfo existingFileInfo = new FileInfo(filePath);
-            FileInfo sophonFileInfo   = new FileInfo(filePath + "_tempSophon");
-
-            // Remove read-only attribute
-            if (existingFileInfo.Exists && existingFileInfo.IsReadOnly)
-            {
-                existingFileInfo.IsReadOnly = false;
-            }
-
-            if (sophonFileInfo.Exists && sophonFileInfo.IsReadOnly)
-            {
-                sophonFileInfo.IsReadOnly = false;
-            }
-
-            // Use "_tempSophon" if file is new or if "_tempSophon" file exist. Otherwise use original file if exist
-            if (!existingFileInfo.Exists || sophonFileInfo.Exists
-                                         || (existingFileInfo.Exists && sophonFileInfo.Exists))
-            {
-                filePath = sophonFileInfo.FullName;
-            }
-
-            // However if the file has already been existed and completely downloaded while _tempSophon is exist,
-            // delete the _tempSophon one to avoid uncompleted files being applied instead.
-            else if (existingFileInfo.Exists && existingFileInfo.Length == asset.AssetSize && sophonFileInfo.Exists)
-            {
-                sophonFileInfo.Delete();
-            }
-
-            await asset.WriteToStreamAsync(
-                                           client,
-                                           () => new FileStream(filePath, FileMode.OpenOrCreate, FileAccess.ReadWrite,
-                                                                FileShare.ReadWrite),
-                                           parallelOptions,
-                                           UpdateSophonFileTotalProgress,
-                                           UpdateSophonFileDownloadProgress,
-                                           UpdateSophonDownloadStatus
-                                          );
-        }
-
         private async ValueTask<int> RunPackageVerificationRoutine(GameInstallPackage asset, CancellationToken token)
         {
             // Reset per size counter
@@ -1899,7 +1211,6 @@ namespace CollapseLauncher.InstallManager.Base
                 WriteAudioLangList(_assetIndex);
             }
         }
-
 
         public virtual async ValueTask<bool> IsPreloadCompleted(CancellationToken token)
         {
@@ -2605,17 +1916,6 @@ namespace CollapseLauncher.InstallManager.Base
                    };
         }
 
-        protected virtual int GetSophonLocaleCodeIndex(SophonData sophonData, string lookupName)
-        {
-            List<string> localeList = sophonData.ManifestIdentityList
-                .Where(x => IsValidLocaleCode(x.MatchingField))
-                .Select(x => x.MatchingField.ToLower())
-                .ToList();
-
-            int index = localeList.IndexOf(lookupName);
-            return Math.Max(0, index);
-        }
-
         protected virtual Dictionary<string, string> GetLanguageDisplayDictFromVoicePackList(
             List<RegionResourceVersion> voicePacks)
         {
@@ -2638,38 +1938,10 @@ namespace CollapseLauncher.InstallManager.Base
             return returnDict;
         }
 
-        protected virtual List<string> GetSophonLanguageDisplayDictFromVoicePackList(SophonData sophonData)
-        {
-            var value = new List<string>();
-            foreach (SophonManifestIdentity identity in sophonData.ManifestIdentityList)
-            {
-                // Check the lang ID and add the translation of the language to the list
-                string localeCode = identity.MatchingField.ToLower();
-                if (IsValidLocaleCode(localeCode))
-                {
-                    string? languageDisplay = GetLanguageDisplayByLocaleCode(localeCode, false);
-                    if (string.IsNullOrEmpty(languageDisplay))
-                    {
-                        continue;
-                    }
-
-                    value.Add(languageDisplay);
-                }
-            }
-
-            return value;
-        }
-
         protected virtual void RearrangeLegacyPackageLocaleOrder(RegionResourceVersion? regionResource)
         {
             // Rearrange the region resource list order based on matching field for the locale
             RearrangeDataListLocaleOrder(regionResource?.voice_packs, x => x.language);
-        }
-
-        protected virtual void RearrangeSophonDataLocaleOrder(SophonData? sophonData)
-        {
-            // Rearrange the sophon data list order based on matching field for the locale
-            RearrangeDataListLocaleOrder(sophonData?.ManifestIdentityList, x => x.MatchingField);
         }
 
         protected virtual void RearrangeDataListLocaleOrder<T>(List<T>? assetDataList, Func<T, string?> matchingFieldPredicate)
@@ -2827,49 +2099,6 @@ namespace CollapseLauncher.InstallManager.Base
                 sw.WriteLine(langString);
             }
         }
-
-        protected virtual void WriteAudioLangListSophon(List<string> sophonVOList)
-        {
-            // Create persistent directory if not exist
-            if (!Directory.Exists(_gameDataPersistentPath))
-            {
-                Directory.CreateDirectory(_gameDataPersistentPath);
-            }
-
-            // If the game does not have audio lang list, then return
-            if (string.IsNullOrEmpty(_gameAudioLangListPathStatic))
-            {
-                return;
-            }
-
-            // Read all the existing list
-            List<string> langList = File.Exists(_gameAudioLangListPathStatic)
-                ? File.ReadAllLines(_gameAudioLangListPathStatic).ToList()
-                : [];
-
-            // Try lookup if there is a new language list, then add it to the list
-            for (int index = 0; index < sophonVOList.Count; index++)
-            {
-                var    packageLocaleCodeString = sophonVOList[index];
-                string langString              = GetLanguageStringByLocaleCode(packageLocaleCodeString);
-                if (!langList.Contains(langString, StringComparer.OrdinalIgnoreCase))
-                {
-                    langList.Add(langString);
-                }
-            }
-
-            // Create the audio lang list file
-            using var sw = new StreamWriter(_gameAudioLangListPathStatic,
-                                            new FileStreamOptions
-                                                { Mode = FileMode.Create, Access = FileAccess.Write });
-            // Iterate the package list
-            foreach (var voIds in langList)
-                // Write the language string as per ID
-            {
-                sw.WriteLine(voIds);
-            }
-        }
-
         #endregion
 
         #region Private Methods - GetInstallationPath
@@ -3643,7 +2872,6 @@ namespace CollapseLauncher.InstallManager.Base
         #endregion
 
         #region Private Methods - StartPackageInstallation
-
         private void MoveFileToIngredientList(List<FilePropertiesRemote> assetIndex, string sourcePath,
                                               string                     targetPath, bool   isSR = false)
         {
